Routines to compute improvements to GP with additional data point
Logic: One point is held out. Then there is a test point. GPs are tuned once without the test point and then a second time with the testpoint. After each tuning, the value of the learnt map at the held out point is computed. The difference in error between 'GP without test point' and 'GP with test point' is the improvement due to the test point.

1. improv_p2h considers 1 map from prms to high res pca.
2. improv_p2l2h considers 2 maps: 1 from prms to low and another from low to high
3. improv_p2l2h_lo is improvement due to adding a low resoution test point
4. improv_p2l2h_hi is improvement due to adding a high resoution test point
5. In either improv_p2l2h_lo/hi, the held out point is a high resolution point.

Main restriction: Only considering first pca components in both low and high.
Need to remove this restriction to allow arbitrary numbers of pcas for low first and then also for high